% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainf_function.R
\name{datainf}
\alias{datainf}
\title{Information from dataframe}
\usage{
datainf(df = WDI(), num = 5)
}
\description{
This function takes in dataframe and outputs the column names, their types, the proportion of missing values and the five most common elements.
}
\examples{
datainf(WDI(),10)
}
